refactor ContextLabel to separate component
