+++ client/lib/ContextLabel/update.js
@@ -1,0 +1,6 @@
+const template = require('./template')
+
+module.exports = function (context) {
+  // Update label.
+  this.component.html(template(context))
+}
