--- client/lib/RootStratum/index.js
@@ -1,2 +1,0 @@
-// TODO a top-level node that displays the first loading sequence.
-// Can be used as a loading animation.
